+++ packages/salesforcedx-vscode/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 59.9.0 - December 13, 2023
+
+## Fixed
+
+#### salesforcedx-vscode-core
+
+- We made some changes under the hood. ([PR #5253](https://github.com/forcedotcom/salesforcedx-vscode/pull/5253))
+  
+#### docs
+
+- We made updates to the Org Browser docs topic. ([PR #5254](https://github.com/forcedotcom/salesforcedx-vscode/pull/5254))
+
 # 59.8.0 - December 6, 2023
 
 ## Added